requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==0.7.16
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncio-throttle==1.0.2
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.12.3
black==25.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
courlan==1.3.2
cryptography==45.0.6
cycler==0.12.1
dateparser==1.2.2
decorator==4.4.2
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
docutils==0.21.2
exceptiongroup==1.3.0
filelock==3.18.0
flake8==7.3.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.7.0
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.107.0
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.29.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
img2pdf==0.6.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
indic-nlp-library==0.92
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.43.0
lxml==5.4.0
lxml_html_clean==0.4.2
macholib==1.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
Morfessor==2.0.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.24.3
oauthlib==3.3.1
openai==1.99.3
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pikepdf==9.9.0
Pillow==10.0.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.1
scikit-learn==1.6.1
scipy==1.11.1
sentence-transformers==5.1.0
shapely==2.0.7
shiboken6==6.9.1
six==1.15.0
sniffio==1.3.1
snowballstemmer==3.0.1
soundfile==0.12.1
soxr==0.5.0.post1
Sphinx==7.4.7
sphinx-argparse==0.4.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.20
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
